merge: merge with replace_xy
