+++ src/tools.py
@@ -1,7 +1,6 @@
 """Tools for checking functions and exporting decompiled program to a .c file"""
 
 # pylint: disable=wrong-import-position, import-error, wrong-import-order
-from math import floor, log2
 import function_code_handling
 import function_handling
 import pyhidra
@@ -9,9 +8,6 @@ import pyhidra
 pyhidra.start()
 from ghidra.app.decompiler import DecompileOptions, DecompInterface
 from ghidra.program.model.data import DataTypeWriter
-
-CONCAT_LEN = 6  # = len("CONCAT")
-BYTE_SIZE = 8
 
 
 def put_program_data_types(program, file_writer, monitor):
@@ -36,49 +32,12 @@ def init_decompiler(program):
     return decompiler
 
 
-def exclude_function(function):
-    """Dumping program data types"""
-    entry_point = function.getEntryPoint()
-    code_unit_at = function.getProgram().getListing().getCodeUnitAt(entry_point)
-    return function_handling.function_in_runtime(function) \
-        or function_handling.function_is_plt(function) \
-        or code_unit_at.getMnemonicString() == "??"
-
-
-def get_nearest_lower_power_2(num):
-    """Rounds a number to nearest lower power of 2"""
-    return 2 ** floor(log2(num))
-
-
-def put_concat(file_writer, code, used_concats):
-    """Puts CONCATXY functions into C code"""
-    concat_cnt = code.count("CONCAT")
-    concat_idx = 0
-    for _ in range(concat_cnt):
-        concat_idx = code.find("CONCAT", concat_idx) + CONCAT_LEN
-        first_size = int(code[concat_idx])
-        second_size = int(code[concat_idx + 1])
-        if (first_size, second_size) in used_concats:
-            continue
-        first_inttype_size = get_nearest_lower_power_2(first_size * BYTE_SIZE)
-        second_inttype_size = get_nearest_lower_power_2(second_size * BYTE_SIZE)
-        concat_name = f"unsigned long CONCAT{first_size}{second_size}"
-        concat_args = f"(uint{first_inttype_size}_t a, uint{second_inttype_size}_t b)\n"
-        concat_body = \
-            f"return ((unsigned long)b) | (unsigned long)a << ({second_size} * {BYTE_SIZE});"
-        concat_signature = concat_name + concat_args
-        concat_function = concat_signature + '{' + '\n' + '\t' + concat_body + '\n' + '}' + '\n'
-        file_writer.println(concat_function)
-        used_concats.add((first_size, second_size))
-    return used_concats
-
-
 def put_functions_signatures(program, file_writer, monitor, decompiler):
     """Puts functions' signatures to C code file"""
     functions_code = []
     single_return_functions = []
     for function in program.getFunctionManager().getFunctions(True):
-        if exclude_function(function):
+        if function_handling.exclude_function(function):
             continue
         results = decompiler.decompileFunction(function, 0, monitor)
         decompiled_function = results.getDecompiledFunction()
@@ -102,8 +61,8 @@ def put_functions_code(functions_code, file_writer, decompiler):
     used_concats = set()
     for function_code in functions_code:
         if "CONCAT" in function_code:
-            used_concats = \
-                put_concat(file_writer, function_code, used_concats)
+            used_concats = function_code_handling.put_concat(file_writer,
+                                                             function_code, used_concats)
         file_writer.println(function_code)
     decompiler.closeProgram()
     decompiler.dispose()
